changed directory parsing method, updated aggregators, and changed reporting
directory parsing now grabs a list of files first, so that it doesn't have to
spawn additional boostrap files just to walk the directory. each file now
runs in its own contained environment. This also makes it easier to do things
like fork later.  Ideally, we could use a web browser to "fork", as with a list
of files, 3-4 tests could be ran at the same time. The phpserializer aggregator
concept could then be ported to JSON to handle this properly.

The second change is in the aggregation method itself. Textaggregator was dropped in favor of phpserializer for simplicity. The way reporting was done was also changed in order to accomodate TAP processing. Instead of determining pass/fail counts on the fly, all reports (including pass reports) are simply passed into
the generateReport method.


git-svn-id: https://snaptest.googlecode.com/svn/trunk@46 85759191-e132-0410-920c-21c98535aa7c