added fatal error handling in sub-process files
this improves serialization handling from the subprocesses, and allows for
fatal errors to be captured. an error state really should be used that
catches everything, to allow for maximum debug during testing.

all reporters take $reports as an array coming into the method, removing
all reliance on the $this keyword in concrete implementations


git-svn-id: https://snaptest.googlecode.com/svn/trunk@47 85759191-e132-0410-920c-21c98535aa7c