added support for requiresInterface to mock objects
some times, somebody needs to clone an existing interface, but to ensure
that LSP is followed, needs to also have the cloned object implement an
interface. This is a prefered method to using requiresInheritance if it is
known that objects will be checking for interfaces as opposed to
abstract class inheritance.


git-svn-id: https://snaptest.googlecode.com/svn/trunk@93 85759191-e132-0410-920c-21c98535aa7c