magic method support added
in PHP, __call can be added, which will cause all manner of problems
in a mocked object.  When __call exists, you can specify the
"requiresMagicMethods" call, which well let any listenTo or setReturnValue
call work. This change updates the mock.php file, and adds small changes for
the way methods are built when they are not listed in public/protected
method arrays


git-svn-id: https://snaptest.googlecode.com/svn/trunk@103 85759191-e132-0410-920c-21c98535aa7c