changed eval statements for parent::parentMethod to use call_user_func_array and improved protected method hasConstructor
originally, the constructor hack was updated by using the protected method
has_constructor. This was replaced with two methods:
  requiresConstructor()
  hasConstructor()
this refactor has no effect on external facing objects.

The second update swaps out the eval statement (which was expensive) for
a call to call_user_func_array($this, 'parent::'.$methodname); which is much
faster. This change would only be noticable on long running tests.