support structure for calling $this->skip() in setUp
after applying the changes for #7, it became clear that setUp() needs to be
able to skip as well. Otherwise, you would have to move all low level
operations into the test, which is not practical.  Supporting this became
a lot bigger than it should.  The net result of the operation IS
* reporters now implement an interface (they should have anyway)
* mock.php understands a mocked constructor in an interface (ignores "parent")
* testcase.stest.php was updated to reflect these changes with a new test
* testcase.php was modified to look at the exception type if thrown in setUp()