parallel processing added to snap handling
the popen() logic from before was moved to a multi-threaded model,
which gives better performance and higher isolation testing. To
improve reuse, a dispatcher was created, which manages the
opening and closing of file handlers, and supplemental classes were
created in order to handle the callbacks the dispatcher needed.
With this, came a large cleanup in the functions space, moving
things to classes where appropriate.  Finally, as many constants as
possible were moved to constants.php