changed TAP output to be streaming
originally, TAP output was compiled at the end. With the new dispatcher
and aggregator code, it was actually really easy to change to a streaming
format. The most substantial change is in the reporter interface, which
now accepts reports for each announce style. This resulted in a small change
to the reporter.php class and testaggregator.php class, as well as having
all the reporters adhear to the new interface.  Finally, snaptest.php was
set up to provide a test count to the test aggregator. The next logical
step is to give the reporter this value as part of generateReport(), although
it is in no way neccessary.

I'm not too happy with how TAP is integrated into the testaggregator. I want to
move those lines out of the testaggregator.php file before making this an
actual commit.